requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.2
autopep8==1.5.6
billiard==3.6.4.0
black==20.8b1
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coverage==5.5
dataclassy==0.7.2
deconstruct==0.6
Django==3.1.7
django-celery-beat==2.2.0
django-filter==2.4.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
fl-flint==0.8.1
flake8==3.9.1
isort==5.8.0
kombu==5.0.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pep8==1.7.1
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
xmltodict==0.12.0
